requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.3.3
exceptiongroup==1.2.1
fastapi==0.105.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
